Experimental C++ Casting

+++ engine.cpp
@@ -116,14 +116,15 @@ char sz_error_check[255];
 edict_t *pfnFindEntityInSphere(edict_t *pEdictStartSearchAfter, const float *org, const float rad) {
    if (debug_engine) {
       fp = UTIL_OpenFoxbotLog();
-      fprintf(fp, "pfnFindEntityInSphere:%p (%f %f %f) %f %d\n", static_cast<void*>(pEdictStartSearchAfter), float((*(Vector*)org).x), float((*(Vector*)org).y), float((*(Vector*)org).z), static_cast<double>(rad), spawn_check_crash_count);
+      fprintf(fp, "pfnFindEntityInSphere:%p (%f %f %f) %f %d\n", static_cast<void *>(pEdictStartSearchAfter), static_cast<double>((*(Vector *)org).x), static_cast<double>((*(Vector *)org).y), static_cast<double>((*(Vector *)org).z),
+              static_cast<double>(rad), spawn_check_crash_count);
 
       if (pEdictStartSearchAfter != nullptr)
          if (pEdictStartSearchAfter->v.classname != 0)
             fprintf(fp, "classname %s\n", STRING(pEdictStartSearchAfter->v.classname));
       fclose(fp);
    }
-   if (spawn_check_crash && rad == 96) {
+   if (spawn_check_crash && static_cast<int>(rad) == 96) {
       spawn_check_crash_count++;
       if (spawn_check_crash_count > 512) {
          // pfnSetSize: 958fd0 (-16.000000 -16.000000 -36.000000) (16.000000 16.000000 36.000000)
@@ -214,12 +215,12 @@ void pfnSetOrigin(edict_t *e, const float *rgflOrigin) {
       for (int bot_index = 0; bot_index < 32; bot_index++) {
          if (bots[bot_index].sentry_edict != nullptr && bots[bot_index].has_sentry) {
             edict_t *pent = e;
-            int l = static_cast<int>(bots[bot_index].sentry_edict->v.origin.z - float((*(Vector*)rgflOrigin).z));
+            int l = static_cast<int>(bots[bot_index].sentry_edict->v.origin.z - static_cast<double>((*(Vector *)rgflOrigin).z));
             if (l < 0)
                l = -l;
 
-            const int xa = static_cast<int>((*(Vector*)rgflOrigin).x);
-            const int ya = static_cast<int>((*(Vector*)rgflOrigin).y);
+            const int xa = static_cast<double>((*(Vector*)rgflOrigin).x);
+            const int ya = static_cast<double>((*(Vector*)rgflOrigin).y);
             const int xb = static_cast<int>(bots[bot_index].sentry_edict->v.origin.x);
             const int yb = static_cast<int>(bots[bot_index].sentry_edict->v.origin.y);
             // FILE *fp;
@@ -258,7 +259,7 @@ void pfnSetOrigin(edict_t *e, const float *rgflOrigin) {
 
    if (debug_engine) {
       fp = UTIL_OpenFoxbotLog();
-      fprintf(fp, "pfnSetOrigin: %p (%f %f %f)\n", static_cast<void*>(e), float((*(Vector*)rgflOrigin).x), float((*(Vector*)rgflOrigin).y), float((*(Vector*)rgflOrigin).z));
+      fprintf(fp, "pfnSetOrigin: %p (%f %f %f)\n", static_cast<void*>(e), static_cast<double>((*(Vector *)rgflOrigin).x), static_cast<double>((*(Vector *)rgflOrigin).y), static_cast<double>((*(Vector *)rgflOrigin).z));
 
       if (e->v.classname != 0)
          fprintf(fp, " name=%s\n", STRING(e->v.classname));
